NEW FEATURE:  Online help
The existing HELP_CHARACTER will now pull up a general help screen when entered as the first character on a line in interactive mode.

The general help screen illustrates how to read the inflection, 'dictionary', and meaning lines.  It contains references to additional screens that describe the codes and abbrevaitions used.