Convert one -i to -ia; demote another -i
A lot of the -i suffixes look questionable.  The only one that has an OED entry was missing:  -ia.

Added -ia (the only -i[x] in OLD)
Demoted one -i (which looks like it may be a typo for -ia) to the end of the list
Moved other -i entries to the end