Mostly V- and L- dictionary updates; new 5 1 adj  "declension"
longiusculus, a, um is a comparative adjective that  behaves like a positive adjective.  The ADJ inflections assume that all comparatives and superlatives have the same ending, so created a new "fifth" declension that copies the 1 1 declension but lists it as comparative.

A side effect is that one of the 4 inflection lookup tables exceeds the maximum entries; increase entries.